Move test helper to flowrlib for reuse across binaries (#770)
Extract the wait_for_then_send() test helper from flowrcli and
flowrgui into a shared generic version in flowrlib::test_helper.
The binary-specific wrappers now delegate to the shared impl,
passing their own ClientMessage::Ack for the handshake.
